所有视频播放功能基本完成了。

+++ app/src/main/cpp/include/DecodeVideoThread.h
@@ -30,8 +30,8 @@ public :
     DecodeVideoThread(AVFormatContext *afc , AVCodecContext  *vc  ,int videoIndex);
     ~DecodeVideoThread();
     void run();
-    int pts ;
-    int apts;
+    int64_t pts ;
+    int64_t apts;
 };
 
 #endif //MYFFMPEG_DECODEVIDEOTHREAD_H
